Ajuste de rotas
ajuste nas rotas de diagnostico
