+++ backend/src/controllers/diagnosticoController.js
@@ -1,7 +1,7 @@
 const connection = require('../configs/connection');
 
 module.exports = {
-    async list(req, res) {
+    async listAluno(req, res) {
 
         await connection('diagnostico_aluno').select('*').then(result => {
             return res.status(200).json(result);
@@ -29,7 +29,7 @@ module.exports = {
             console.log(result)
             let diagnosticos = [];
             result.map( diagnostico => {
-                let index = diagnosticos.findIndex( e => e.id == diagnostico.idDiagnostico)
+                let index = diagnosticos.findIndex( e => e.id == diagnostico.idAluno)
                 if ( index == -1){
                     diagnosticos.push({
                         id: diagnostico.idDiagnostico,
@@ -67,7 +67,7 @@ module.exports = {
 
     },
 
-    async create(req, res) {
+    async createAluno(req, res) {
 
         const { idAluno, idIndice, idPeriodo } = req.body;
 
@@ -80,5 +80,43 @@ module.exports = {
         }).catch(error => {
             return res.status(500).json(error);
         });
+    },
+
+
+//////////////Turma
+
+    async listTurma(req, res) {
+
+        await connection('diagnosticar_turma').select('*').then(result => {
+            return res.status(200).json(result);
+        }).catch(error => {
+            return res.status(500).json(error);
+        });
+    },
+
+    async listPorPeriodo(req, res) {
+        const { idPeriodo } = req.params;
+
+        await connection('diagnosticar_turma').select('*').where('idPeriodo', idPeriodo || 0).then(result => {
+            return res.status(200).json(result);
+        }).catch(error => {
+            return res.status(500).json(error);
+        });
+
+    },
+
+    async createTurma(req, res) {
+
+        const { idPerfil_turma, idPeriodo } = req.body;
+
+        await connection('diagnosticar_turma').insert({
+            idPerfil_turma,
+            idPeriodo
+        }).then(result => {
+            return res.status(200).json(result);
+        }).catch(error => {
+            return res.status(500).json(error);
+        });
     }
+
 }
